+++ semana10/lubax/src/router/goToPages.js
@@ -1,0 +1,11 @@
+export const goBack = (history) => {
+  history.goBack();
+};
+
+export const goToLoginPage = (history) => {
+  history.push("/login");
+};
+
+export const goToTripPage = (history) => {
+  history.push("/trips/list");
+};
